In Config Dashboard do not watch Device Changes directly - too resource intensive
If we watch the Device changes in Config Dashboard, it will open up a stream connection to
server for each device - above 10 devices this becomes unfeasible because it puts too much
of a burden on the onos-config pod

Instead display the list of device changes known from the network config. This will give
us access to their paths and values. When we view individual devices in the Config View
we will watch the Device Changes there for an individual device